+++ docs/Changelog.md
@@ -4,6 +4,7 @@ title: Changelog
 
 ## Not released yet
 
+* 2018-05-29 Remerge command to support free remerges for already merged posts
 * 2018-05-24 Fetch functionality correctly handles 8bit encodings
 * 2018-05-17 Merge functionality
 * 2018-05-10 Fetch functionality stores sync timestamps for every item for later